requirements.txt
streamlit
pandas
matplotlib
langchain
langchain-core
langchain-google-genai
google-generativeai
